Sikkim chatbot/requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.1
transformers==4.36.0
accelerate==0.25.0
bitsandbytes==0.41.3
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0
aiofiles==23.2.0 